backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP requests
requests==2.31.0

# Audio processing and transcription
openai-whisper==20231117
torch==2.1.1
numpy==1.24.3

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
